Referential jointing.
Term 'referential joint', not 'waylink'.  The implied meaning
of 'waylink' is too general.  It covers all connections
that involve the way.  It conflicts with other terms.

Terms 'jointer' and 'subjoining', not 'source' and 'target'.
The latter fit poorly a relation whose orientation often varies
in direction (or goes directionless) according to how one views it.